+++ packages/@dcl/sdk-commands/src/commands/start/data-layer/fs.ts
@@ -1,10 +1,15 @@
 import { FileSystemInterface } from '@dcl/inspector'
-import fs from 'fs-extra'
+import { IFileSystemComponent } from '../../../components/fs'
 
-export function createFsFromNode(): FileSystemInterface {
+export function createFsFromFsComponent(fs: IFileSystemComponent): FileSystemInterface {
   return {
     async existFile(filePath: string): Promise<boolean> {
-      return fs.existsSync(filePath)
+      try {
+        await fs.access(filePath, fs.constants.F_OK | fs.constants.R_OK)
+        return true
+      } catch (error) {
+        return false
+      }
     },
     async readFile(filePath: string): Promise<Buffer> {
       return fs.readFile(filePath)
@@ -13,11 +18,13 @@ export function createFsFromNode(): FileSystemInterface {
       await fs.writeFile(filePath, content)
     },
     async readdir(dirPath: string): Promise<{ name: string; isDirectory: boolean }[]> {
-      const result = await fs.readdir(dirPath, { withFileTypes: true })
-      return result.map((item) => ({
-        name: item.name,
-        isDirectory: item.isDirectory()
-      }))
+      const result = await fs.readdir(dirPath)
+      return Promise.all(
+        result.map(async (name) => ({
+          name: name,
+          isDirectory: await fs.directoryExists(name)
+        }))
+      )
     }
   }
 }
