+++ packages/@dcl/sdk-commands/src/commands/start/data-layer/rpc.ts
@@ -3,7 +3,7 @@ import { createRpcServer, RpcServer, RpcServerPort } from '@dcl/rpc'
 
 import * as codegen from '@dcl/rpc/dist/codegen'
 import { CliComponents } from '../../../components'
-import { createFsFromNode } from './fs'
+import { createFsFromFsComponent } from './fs'
 
 export type DataLayerContext = {
   dataLayerHost: DataLayerHost
@@ -14,13 +14,12 @@ export type DataLayer = {
   context: DataLayerContext
 }
 
-export async function createDataLayer({ fs: _fs }: Pick<CliComponents, 'fs'>): Promise<DataLayer> {
-  // TODO: implement createFsFromIFileSystemComponent(fs)
-  const dataLayerHost = await createDataLayerHost(createFsFromNode())
+export async function createDataLayer({ fs, logger }: Pick<CliComponents, 'fs' | 'logger'>): Promise<DataLayer> {
+  const dataLayerHost = await createDataLayerHost(createFsFromFsComponent(fs))
   const context: DataLayerContext = {
     dataLayerHost
   }
-  const rpcServer = createRpcServer<DataLayerContext>({})
+  const rpcServer = createRpcServer<DataLayerContext>({ logger })
 
   async function rpcHandler(serverPort: RpcServerPort<DataLayerContext>) {
     // TODO: dataLayer as any
