+++ packages/@dcl/sdk-commands/src/commands/start/data-layer/ws.ts
@@ -3,7 +3,7 @@ import { WebSocket } from 'ws'
 import { PreviewComponents } from '../types'
 import { DataLayer } from './rpc'
 
-// TODO: dataLayer should be an optional component
+// TODO: dataLayer should be an optional component after WKC supports it
 export async function handleDataLayerWs(components: PreviewComponents, ws: WebSocket, dataLayer: DataLayer) {
   const wsTransport = WebSocketTransport(ws as any)
   dataLayer.rpcServer.attachTransport(wsTransport, dataLayer.context)
