+++ packages/@dcl/sdk-commands/src/logic/catalyst-requests.ts
@@ -1,7 +1,31 @@
 import { Entity } from '@dcl/schemas'
-import { fetch } from 'undici'
+import { CliComponents } from '../components'
+
+export type DAOCatalyst = {
+  baseUrl: string
+  owner: string
+  id: string
+}
+
+type CatalystInfo = {
+  url: string
+  timestamp: number
+  entityId: string
+}
+
+export type Network = 'mainnet' | 'goerli'
+
+export async function daoCatalysts(
+  { fetch }: Pick<CliComponents, 'fetch'>,
+  network: Network = 'mainnet'
+): Promise<Array<DAOCatalyst>> {
+  const tld = network === 'mainnet' ? 'org' : 'zone'
+  const resp = await (await fetch.fetch(`https://peer.decentraland.${tld}/lambdas/contracts/servers`)).json()
+  return resp as DAOCatalyst[]
+}
 
 export async function fetchEntityByPointer(
+  { fetch }: Pick<CliComponents, 'fetch'>,
   baseUrl: string,
   pointers: string[]
 ): Promise<{
@@ -16,7 +40,7 @@ export async function fetchEntityByPointer(
 
   const activeEntities = baseUrl + '/content/entities/active'
 
-  const response = await fetch(activeEntities, {
+  const response = await fetch.fetch(activeEntities, {
     method: 'post',
     headers: { 'content-type': 'application/json', connection: 'close' },
     body: JSON.stringify({ pointers })
@@ -28,4 +52,27 @@ export async function fetchEntityByPointer(
     baseUrl,
     deployments
   }
+}
+
+export async function getPointers(
+  components: Pick<CliComponents, 'fetch' | 'logger'>,
+  pointer: string,
+  network: Network = 'mainnet'
+) {
+  const catalysts = await daoCatalysts(components, network)
+  const catalystInfo: CatalystInfo[] = []
+
+  for (const { baseUrl } of catalysts) {
+    try {
+      const result = await fetchEntityByPointer(components, baseUrl, [pointer])
+      const timestamp = result.deployments[0]?.timestamp
+      const entityId = result.deployments[0]?.id || ''
+
+      catalystInfo.push({ timestamp, entityId, url: baseUrl })
+    } catch (err: any) {
+      components.logger.log('Error fetching catalyst pointers', err)
+    }
+  }
+
+  return catalystInfo
 }
